--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -5,13 +5,17 @@ Version: 0.1.2
 Authors@R: c(
   person("Evan", "Flack", email = "evanjflack@gmail.com",
          role = c("aut", "cre")),
-  person("Edward", "Jee", role = c("aut"))
+  person("Edward", "Jee", role = c("aut")),
+  person("Kyle", "Butts", role = c("ctb"), email = "kyle.butts@colorado.edu",
+         comment = c(ORCID = "0000-0002-9048-8059"))
   )
 Maintainer: Evan Flack <evanjflack@gmail.com>
 Description: Decomposition for differences-in-differences with variation in
     treatment timing from Goodman-Bacon (2018) <doi:10.3386/w25018>.
 License: MIT + file LICENSE
-Depends: R (>= 2.10)
+Depends: 
+    fixest,
+    R (>= 2.10)
 Suggests:
     knitr,
     rmarkdown,