--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -8,7 +8,7 @@ Authors@R: c(
   person("Edward", "Jee", role = c("aut")),
   person("Kyle", "Butts", role = c("ctb"), email = "kyle.butts@colorado.edu",
          comment = c(ORCID = "0000-0002-9048-8059")),
-  person("Zachary", "Porreca", role = c("ctb"), email = "zjp00003@mix.wvu.edu"),
+  person("Zachary", "Porreca", role = c("ctb"), email = "zjp00003@mix.wvu.edu")
   )
 Maintainer: Evan Flack <evanjflack@gmail.com>
 Description: Decomposition for differences-in-differences with variation in