--- a/NEWS.md
+++ b/NEWS.md
@@ -1,11 +1,16 @@
 # bacondecomp 0.1.3
-This version uses fixest to greatly increase the speed.
+
+- This version uses fixest to greatly increase the speed.
+- Export bacon_summary to get aggregates for three kinds of comparisons: Earlier vs. Later (Good), Treated vs. Untreated (Good), and Later vs. Earlier (Bad)
 
 # bacondecomp 0.1.2
-This version removes an unnecessary balanced panel check/error.
+
+- This version removes an unnecessary balanced panel check/error.
 
 # bacondecomp 0.1.1
-This version fixes the summary print from bacon(), to match the STATA package output.
+
+- This version fixes the summary print from bacon(), to match the STATA package output.
 
 # bacondecomp 0.1.0
-First release!
+
+- First release!