For a speed comparison:

```{r}
# A tibble: 2 × 13
  expression      min   median `itr/sec` mem_alloc `gc/sec` n_itr  n_gc total_time result       memory                 time           gc              
  <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl> <int> <dbl>   <bch:tm> <list>       <list>                 <list>         <list>          
1 fixest      143.1ms  158.7ms    6.26      80.3MB    2.09      3     1   479.28ms <df [4 × 5]> <Rprofmem [1,325 × 3]> <bench_tm [3]> <tibble [3 × 3]>
2 orig          31.8s    32.2s    0.0307   872.7MB    0.123     3    12      1.63m <df [4 × 5]> <Rprofmem [1,329 × 3]> <bench_tm [3]> <tibble [3 × 3]>
```