@kylebutts 2021-09-14
I did three more things:

1. Add controls version
2. Export `bacon_summary` which is just `print_summary` but now with a more user-facing name
3. Update README to show that as well