@kylebutts 2021-09-14
hey! I'm not sure why R-CMD-check is failing. I think the github action is not working because pandoc is not installed, so maybe you have to update your action?